requirements.txt
Pandas >= 2.0
regex == 2023.12.25
biopython == 1.81
